refactor: split extension.test.ts into 6 focused test files
Split the 4,761-line test file into smaller, domain-focused files to reduce
conflicts and improve maintainability:

- testSetup.ts: Shared setup/teardown utilities
- session.test.ts: Session management tests (~1,600 lines)
- gitChanges.test.ts: Git changes panel and diff tests (~1,200 lines)
- configuration.test.ts: Config, paths, and storage tests (~1,000 lines)
- projectManager.test.ts: Project Manager service tests (~370 lines)
- edgeCases.test.ts: Edge cases and activation tests (~660 lines)

All 468 tests pass. Test runner auto-discovers new files via existing
.vscode-test.mjs glob pattern.
